Add ability to change grid size
Functionality added to change grid size button, and a number
between 1-100 can be input to create a grid of 1-100 x 1-100.
This provides the artist with more cells to use for their
drawings - the total width/height of the grid is not changed
only the 'resolution'.
This has involved adding a function to delete the current grid.
A new grid is then created and event listeners re-created for
all cells.