Add mobile compatibility and better responsiveness
Drawing can now be completed with a mobile device by pressing
and holding the finger over the grid. This includes an event
listener for "touchmove" and tracking the cells with
"elementFromPoint".
Responsiveness of the grid has been improved by setting a
max-width for cells and row-wrapper, and scaling by use of
'flex'. Flex-wrap has also been added for the parent page
wrapper, so the options box is wrapped underneath the grid
when the width of the viewport is reduced.