Add option to completely clear the Etchasketch
Add button to options menu which clears the current Etchasketch.
Function clearGrid created for this purpose. This gives the user
a convenient option to remove all current 'art' on the Etcha-
sketch. The function creates a new grid of the same size as
the current grid on the page.
Also added flex column styling to the options menu for aesthetics.