Add rainbow colour mode
Ability to change the colour mode in options menu added. The
default of black can be toggled to rainbow mode, which colours
the cells with a random colour. The changeCell function has
been updated to include the colour mode as an input parameter.
A function has been created to return a random number between
0-255, which is applied to designate the random colour in rgb
format (a random number is generated for each of red, green and
blue). The random colour is applied by directly manipulating
the style of the targetted cell.