Fix eraser bug which prevented re-colouring cells
When colouring a cell black and erasing, the cell could not be
re-coloured black. This was due to the black colouring method
using a class to apply the colour, while eraser and rainbow modes
use in-line styling. The class method for applying black colour
has been removed and replaced with in-line styling.
'justify-content' for page flex wrapper has been changed to center
for easier access on desktop.