Add opacity mode (progressive darkening effect)
Opacity mode can be activated by a button added to the menu. When
active, the mode will progressively darken the colour of the cell
by an opacity of 10% for each click. The "changeCell" function has
been modified to include code for this effect.
Toggle buttons for eraser and opacity modes are now coloured red or
green depending if they are switched off or on for better presenta-
tion.
Background colour of grid set to white, otherwise the colour paint-
ed in opacity mode is impacted by the page background.
Also added around border to Etch-a sketch pad, and modified the
flex-basis to ensure the additional width of the element due to
border is accounted for.