Fixed button colours, clearing grid & eraser mode
1. On mobile devices, the options menu button colours were being
stuck in the hover styling. Added media query to hover styling to
fix.
2. Clearing the grid was not creating a new grid due to a name
change of a div element which was not reflected in a selector in
the deletedGrid function (to append the new row wrapper).
3. Eraser mode was not working following the previous commit.
The changeCell function had been updated to use "remove attribute"
to remove styling instead of styling the cell white. However, this
was incorrectly applied to "background-colour" rather than "style"
itself.